requirements.txt
Pdfminer.six
